// src/pages/admin/modules/reports/wasteCollectedSummary/types.ts
export type ApiRow = {
  date: string;
  total_vehicle?: number | string;
  vehicle_count?: number | string;
  total_vehicle_count?: number | string;
  vehicles?: number | string;
  no_of_vehicle?: number | string;
  no_of_vehicles?: number | string;
  total_trip?: number | string;
  dry_weight: number;
  wet_weight: number;
  mix_weight: number;
  total_net_weight: number;
  average_weight_per_trip: number;
};